Add default Clojure dependency to release profile
Enabling the 'sonatype-oss-release' profile disables any
<activeByDefault>true</activeByDefault> profiles. See
http://maven.apache.org/guides/introduction/introduction-to-profiles.html

Our 'default' profile is what adds the Clojure dependency. We need
that profile enabled during the release:prepare step to compile and
test.

For some reason, profiles passed on the mvn command line in Hudson are
not propagating to the nested mvn execution done by release:prepare.
This change should work around that, in addition to being more robust
in general.